Made a saturation EffectSaturate.fx while testing how to make the glow one. Made a Bloom class who stores the EffectGlow.fx that causes the bloom glowing effect on a sprite.
