+++ MGCore/GraphicsTestRig.cs
@@ -51,7 +51,7 @@ namespace MGCore
             base.LoadContent();   //mabye dont do this.. conetn get load in each test.. and w em might unlaod it too... or IDispose it...  //TODO
             
 
-            CurrentDrawTest=new Invert();
+            CurrentDrawTest=new Bloom();
             
             CurrentDrawTest.Initialize(GraphicsDevice, GraphicsDeviceManager, Content);
